Mostly working D3 circle graph of family structure working.  Mouseover still has bugs due to 'this' problem. Spent a bunch of time getting Family/Person to json and back, then found out that passing a Kotlin class to D3 as data does not work.  must revert to just JSON.parse and pass in a JSONObject.  blah. Ported JS code to Kotlin, fully working with above exception about 'this', in js this would be the node that performed the callback, which doesn't appear to be available in kotlin.  May partly be D3 overshadowing.

+++ LifeCommon/src/main/kotlin/com/willhughes/life/LifeRunner.kt
@@ -28,7 +28,7 @@ object LifeRunner {
             runLife()
             if (WorldState.lives.size == 0) {
                 clearInterval(timerId)
-            } else if (currentGeneration as Int > ConfigReader.jsonObj.initialRun as Int) {
+            } else if (currentGeneration > ConfigReader.jsonObj.initialRun as Int) {
                 Logger.noLogging = false
                 clearInterval(timerId)
                 timerId = setInterval(runTimer(), DELAY)
